SimpleViewer: Move shader ball loading to utils
